#92 Move examples to it's own file
