+++ .github/workflows/docker-build-container.yaml
@@ -1,22 +1,18 @@
 name: docker-build-container.yaml
+
 on:
   pull_request:
     branches:
       - main
   workflow_dispatch:
+
 jobs:
   docker-build-container:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
+      - name: Build docker image
+        uses: Senzing/github-action-docker-buildx-build@latest
         with:
-          fetch-depth: "0"
-      - name: git config
-        run: |
-          git config user.name github-actions
-          git config user.email support@senzing.com
-      - name: Git Update submodules
-        run: git submodule update --init --recursive
-      - name: Build Docker container
-        run: make docker-build
+          image-repository: senzing/test-ground
+          password: ${{ secrets.DOCKERHUB_PASSWORD }}
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
