+++ .github/workflows/docker-push-containers-to-dockerhub.yaml
@@ -9,24 +9,11 @@ jobs:
   docker-push-containers-to-dockerhub:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
+      - name: Build docker image and push to DockerHub
+        uses: Senzing/github-action-docker-buildx-build@latest
         with:
-          fetch-depth: "0"
-      - name: Set up QEMU
-        uses: docker/setup-qemu-action@v3
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v3
-      - name: Docker list platforms
-        run: docker buildx ls
-      - name: Dockerhub login
-        run: echo "${{ secrets.DOCKERHUB_PASSWORD }}" | docker login -u "${{ secrets.DOCKERHUB_USERNAME }}" --password-stdin
-      - name: Docker build images
-        run: |
-          docker buildx build \
-            --no-cache \
-            --platform linux/arm64,linux/amd64 \
-            --push \
-            --tag senzing/template-go:${{ github.ref_name }} \
-            --tag senzing/template-go:latest \
-            .
+          build-options: "--push"
+          image-repository: senzing/test-ground
+          image-tag: ${{ github.ref_name }}
+          password: ${{ secrets.DOCKERHUB_PASSWORD }}
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
